auralink-communication-service/api/requirements.txt
# AuraLink Communication Service API Dependencies

# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client
httpx==0.26.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Redis
redis==5.0.1
hiredis==2.3.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring
prometheus-client==0.19.0

# Logging
python-json-logger==2.0.7

# Environment
python-dotenv==1.0.0
